Specification directory structure, name changed to lowercase
