+++ cita-monitor-server/config/alert.rules
