+++ cita-monitor-server/config/dashboards/CITANodeInfoDashboard.json
