+++ cita-monitor-server/config/dashboards/HostInfoDashboard.json
