+++ cita-monitor-server/config/dashboards/RabbitMQDashboard.json
